Add distributed sync
Enable name server to compare timestamps if available across notify disconnects to prevent removing a server from a group if the chat server register occured after the notify. disconnect message was constructed. This allows the server to recover from a crash and can register before a client prevents a disconnect